refactor(dashboard-ui): share the theme module with both dashboards
The preference store, the topbar picker and the pre-paint init script were
web-ui-local, but nothing in them is host-specific: they read localStorage
and matchMedia and write document.documentElement, all lazily inside
functions. Only the INJECTION POINT of THEME_INIT_SCRIPT differs — a Next
root layout here, a Vite index.html elsewhere — and that is a line in each
host, not a fork of the module.

Moved rather than copied. A second copy of a preference store, a media-query
subscription and a pre-paint script is a copy that drifts, and dashboard-ui
is already where shared presentational composites live. ThemeToggle was
built to qualify: it inlines its own SVGs, takes its primitives from ui-kit,
and carries 'use client'.

THEME_INIT_SCRIPT stays a flat literal — it is injected with
dangerouslySetInnerHTML, so interpolating the constants into it would be
code construction, and JSON.stringify does not escape </script>. Both suites
move with it, because what pins that literal against the constants has to
keep running wherever the literal lives.

Note the header now records why the preference stays browser-local even for
a host that HAS accounts: a theme belongs to the screen you are looking at,
not to who you are.
